backend/pyproject.toml
[tool.poetry]
name = "iot-project"
version = "0.1.0"
description = ""
authors = ["Dawid Jakóbczak <djakobczak.student@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
dash = "^2.0.0"
plotly = "^5.3.1"
SQLAlchemy = "^1.4.26"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
